pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "langgraph_poc"
version = "0.1.0"
description = "LangGraph-based agent proof of concept"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  {name = "LangGraph POC Team"}
]
dependencies = [
  "langgraph>=0.2.10",
  "langchain>=0.2.9",
  "langchain-core>=0.2.24",
  "langchain-community>=0.2.11",
  "langchain-openai>=0.1.15",
  "langchain-text-splitters>=0.0.2",
  "langsmith>=0.1.77",
  "qdrant-client>=1.9.0",
  "chromadb>=0.5.3",
  "pyyaml>=6.0.1",
  "pydantic>=2.8.2",
  "python-dotenv>=1.0.1",
  "tqdm>=4.66.4",
  "typer>=0.12.3",
  "rich>=13.7.1",
  "networkx>=3.3",
  "aiohttp>=3.9.5",
  "tenacity>=8.4.2",
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5.6",
  "mypy>=1.10.1",
  "pytest>=8.2.2"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
